pyproject.toml
[project]
name = "circle-packing-grid"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "psycopg2-binary>=2.9.9",
  "shapely>=2.0.2",
  "geopandas>=0.14.1",
  "numpy>=1.24.0",
  "scipy>=1.11.0",
  "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
